pyproject.toml
[tool.poetry]
name = "sample-python-tests"
version = "0.1.0"
description = "Sample Python project built with Poetry for automated tests"
authors = ["Fabio Bressler <fabio.bressler@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = ""
repository = "https://github.com/xyz/python-poetry-sample/"
documentation = "https://xyz.com/python-poetry/"
keywords = [
    "Poetry",
    "Virtual Environments",
    "Tutorial",
    "Packages",
    "Packaging"
]
packages = [
  { include = "app", from="." }
]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]
pytest = "*"
pytest-cov = "*"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "app.main:main"